lambdas/qr-detector/requirements.txt
botocore<1.35.0
boto3==1.34.64
Pillow==10.3.0
pdf2image==1.17.0
opencv-python-headless==4.8.1.78
numpy==1.26.4

